requirements.txt
pydantic
pydantic-settings
bson
pymongo
python-dotenv
# docker
requests
# pytz # timezone

fastapi
uvicorn[standard]
redis

# # FastAPI form data
# python-multipart

# data
pandas

# openai/GPT API
openai